EXTRACTED FUNCTIONS FROM actions.py TO events.py
