+++ zerver/views/events_register.py
@@ -4,7 +4,7 @@ from django.http import HttpRequest, HttpResponse
 from typing import Text
 from typing import Iterable, Optional, Sequence
 
-from zerver.lib.actions import do_events_register
+from zerver.lib.events import do_events_register
 from zerver.lib.request import REQ, has_request_variables
 from zerver.lib.response import json_success
 from zerver.lib.validator import check_string, check_list, check_bool
